renome le fichier et ajoute la mesure de l'angle pitch
